Add files via upload
Hey, guys! This is a website that looks like Starbucks. this is clone website It was made with front-end languages like HTML and CSS.
This is just the landing page of a Starbucks clone, and I'm still working on the other pages, such as the sign-in and sign-up pages, as well as the pages that will link you to the other pages on this clone website. This clone webpage was developed for practicing Web-development Only**